some stuff edited

+++ bonus/srcs/parsing/verif_map.c
@@ -6,7 +6,7 @@
 /*   By: ulayus <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/11 13:05:59 by ulayus            #+#    #+#             */
-/*   Updated: 2023/04/24 13:22:06 by ulayus           ###   ########.fr       */
+/*   Updated: 2023/04/24 13:33:13 by ulayus           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ static bool	check_char_context(const char **map, int i, int j)
 		&& ft_isspace(map[i][j]) == false && map[i][j] != '\n')
 		valid = false;
 	else if (ft_isspace(map[i][j]) == true && map[i + 1] != NULL
-		&& ft_strlen(map[i + 1]) > j
+		&& ft_strlen(map[i + 1]) > (size_t)j
 		&& (map[i + 1][j] == '0' || map[i + 1][j] == 'D'))
 	{
 		if (map[i + 1])
